# Conflicts:
#	DotaStats.xcodeproj/project.pbxproj
#	DotaStats/AppCoordinator.swift

+++ DotaStats/Data/Networking/NetworkingModels/MatchDetail.swift
@@ -130,7 +130,7 @@ struct MatchDetail: Decodable {
         let lose: Int?
         let totalGold: Int?
         let totalXp: Int?
-        let killsPerMin: Int?
+        let killsPerMin: Double?
         let kda: Int?
         let abandons: Int?
         let neutralKills: Int?
